vana_multi_agent/requirements.txt
# VANA Multi-Agent System Dependencies - ADK Compatible

# Google ADK and AI Platform (Matching official ADK samples)
google-adk>=1.0.0
google-cloud-aiplatform[adk,agent_engines]>=1.93.0
google-genai>=1.14.0

# Core Dependencies (ADK Compatible)
python-dotenv>=1.0.0
uvicorn>=0.34.0
fastapi>=0.115.0
pydantic>=2.7.0,<3.0.0

# Enhanced Tool Dependencies
requests>=2.31.0
numpy>=1.24.0
pandas>=2.1.0

# Vector Search Dependencies
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Knowledge Graph Dependencies
networkx>=3.2.0
spacy>=3.7.0

# Web Search Dependencies
google-api-python-client>=2.100.0

# Monitoring and Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Dashboard Dependencies (for admin portal)
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.2.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
